под шкалой не меняется значок рубль/доллар
единицы измерения когда доллар должны быть 0.01-0.12, а не 1-12

+++ app/src/main/java/org/sigmapool/sigmapool/provider/currency/models/CurrencyParams.kt
@@ -3,8 +3,8 @@ package org.sigmapool.sigmapool.provider.currency.models
 import androidx.annotation.ArrayRes
 
 class CurrencyParams(
-    val scaleFrom: Int,
-    val scaleTo: Int,
+    val scaleFrom: Float,
+    val scaleTo: Float,
     val initValue: Int,
     @ArrayRes val stringArrayRes: Int,
     val step: Int,
